rename s.i.yaml to s.i.project_loader.
