+++ snapcraft/internal/project_loader.py
